move entity types to entity_type.py and handle custom errors
