+++ iaso/api/mobile/entity.py
@@ -115,60 +115,3 @@ class MobileEntityViewSet(ModelViewSet):
         return queryset
 
 
-class MobileEntityTypeSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = EntityType
-        fields = [
-            "name",
-            "created_at",
-            "updated_at",
-            "reference_form",
-            "account",
-            "is_active",
-        ]
-
-    created_at = TimestampField()
-    updated_at = TimestampField()
-
-
-class MobileEntityTypesViewSet(ModelViewSet):
-    results_key = "entitytypes"
-    remove_results_key_if_paginated = True
-    filter_backends = [filters.OrderingFilter, DjangoFilterBackend]
-    pagination_class = LargeResultsSetPagination
-
-    def get_serializer_class(self):
-        return MobileEntityTypeSerializer
-
-    def get_queryset(self):
-
-        queryset = EntityType.objects.filter(account=self.request.user.iaso_profile.account)
-
-        return queryset
-
-    @action(detail=False, methods=["get"], url_path=r"(?P<type_pk>\d+)/entities")
-    def get_entities_by_types(self, *args, **kwargs):
-        limit_date = self.request.query_params.get("limit_date", None)
-        json_content = self.request.query_params.get("json_content", None)
-        type_pk = self.request.parser_context.get("kwargs")["type_pk"]
-        p = Prefetch(
-            "instances",
-            queryset=Instance.objects.filter(deleted=False).exclude(file=""),
-            to_attr="non_deleted_instances",
-        )
-
-        queryset = Entity.objects.filter(account=self.request.user.iaso_profile.account, entity_type__pk=type_pk)
-        queryset = queryset.prefetch_related(p).prefetch_related("non_deleted_instances__form")
-
-        if limit_date:
-            queryset = queryset.filter(instances__updated_at__gte=limit_date)
-
-        if json_content:
-            q = jsonlogic_to_q(jsonlogic=json_content, field_prefix="json__")  # type: ignore
-            queryset = queryset.filter(q)
-
-        page = self.paginate_queryset(queryset)
-
-        serializer = MobileEntitySerializer(page, many=True)
-
-        return self.get_paginated_response(serializer.data)
